requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
chromadb
langchain
langchain-community
transformers
torch
sentence-transformers
pandas
numpy
pydantic[email]
python-dotenv
httpx
pytest
pytest-asyncio
